requirements.txt
openai
phidata
python-dotenv
yfinance
packaging
duckduckgo-search
fastapi
uvicorn
streamlit